+++ src/lambdas/metrics_updater/handler.py
@@ -2,6 +2,7 @@ from http import HTTPStatus
 
 from modular_sdk.commons.trace_helper import tracer_decorator
 
+from helpers.constants import ServiceOperationType
 from helpers.lambda_response import (
     SREException,
     MetricsUpdateException,
@@ -22,7 +23,10 @@ class MetricsUpdater(EventProcessorLambdaHandler):
     def build(cls) -> 'MetricsUpdater':
         return cls()
 
-    @tracer_decorator(is_job=True, component='metrics')
+    @tracer_decorator(
+        is_job=True, 
+        component=ServiceOperationType.UPDATE_METRICS.value,
+    )
     def handle_request(self, event, context):
         # todo validate event
         dt = event.get('data_type')